Adding testcases to autoinstrumentation
Factoring out the `isPodEligible` function to be able to test this
as a separate piece from injection since these tests only cover
`extractLibInfo`.

It's important to know the difference between the two... the second
one only gets called _after_ the first and the test cases cover
particular outcomes which might not happen because of the
eligibility check at the beginning of `Webhook.inject` --
now also being tested here.

We can see in the added test cases that disabling mutate_unlabelled
changes the libs that are output and is generally congruent with
eligibility.

However! in the case of `all-lib.version` -- expected to add all
of the latest libraries, it will only work if the feature flag if
mutate_ulabelled is on and instrumentation is off.