Decoupling the webhooks
Cleans up some of the eligibility checks for getting library
infromation in auto_instrumentation.

In doing so it decouples webhook global dependencies.

It moves the shared global into its own file, and it can be later
used to be moved into its own package. I kept the namespace filter
in the auto_instrumentation package since it relies on apm configs.